requirements.txt
torch
pyro-ppl
pytest